Updated jars. Migrated native code back from FIPS project.

+++ native/intel/cfb/AesCFBEncrypt.cpp
@@ -92,7 +92,7 @@ namespace intel {
             }
 
 
-            return dest - destStart;
+            return (size_t)(dest - destStart);
         }
